Port PMaybeData utils into -api
