+++ plutarch-api/src/Plutarch/Api.hs
@@ -1,13 +1,14 @@
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RoleAnnotations #-}
+{-# LANGUAGE ViewPatterns #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 {- | = Note
 
-The 'Value.PValue' and 'AssocMap.PMap'-related functionality can be found in
-other modules, as these clash with the Plutarch prelude. These should be
-imported qualified.
+The 'Value.PValue', 'AssocMap.PMap' and 'Interval.PInterval'-related
+functionality can be found in other modules, as these clash with the
+Plutarch prelude. These should be imported qualified.
 -}
 module Plutarch.Api (
   -- * Contexts
@@ -85,6 +86,15 @@ module Plutarch.Api (
 
   -- ** Types
   PMaybeData (..),
+
+  -- ** Utilities
+  pfromDJust,
+  pisDJust,
+  pmaybeData,
+  pdjust,
+  pdnothing,
+  pmaybeToMaybeData,
+  passertPDJust,
 ) where
 
 import Codec.Serialise (serialise)
@@ -1117,6 +1127,95 @@ pparseDatum = phoistAcyclic $ plam $ \dh datums ->
   pmatch (AssocMap.plookup # dh # datums) $ \case
     PNothing -> pcon PNothing
     PJust datum -> pcon . PJust $ ptryFrom (pto datum) fst
+
+{- | Extracts the element out of a 'PDJust' and throws an error if its
+argument is 'PDNothing'.
+
+@since 2.1.1
+-}
+pfromDJust ::
+  forall (a :: PType) (s :: S).
+  PIsData a =>
+  Term s (PMaybeData a :--> a)
+pfromDJust = phoistAcyclic $
+  plam $ \t -> pmatch t $ \case
+    PDNothing _ -> ptraceError "pfromDJust: found PDNothing"
+    PDJust x -> pfromData $ pfield @"_0" # x
+
+{- | Yield 'PTrue' if a given 'PMaybeData' is of the form @'PDJust' _@.
+
+@since 2.1.1
+-}
+pisDJust ::
+  forall (a :: PType) (s :: S).
+  Term s (PMaybeData a :--> PBool)
+pisDJust = phoistAcyclic $
+  plam $ \x -> pmatch x $ \case
+    PDJust _ -> pconstant True
+    _ -> pconstant False
+
+{- | Special version of 'pmaybe' that works with 'PMaybeData'.
+
+@since 2.1.1
+-}
+pmaybeData ::
+  forall (a :: PType) (b :: PType) (s :: S).
+  PIsData a =>
+  Term s (b :--> (a :--> b) :--> PMaybeData a :--> b)
+pmaybeData = phoistAcyclic $
+  plam $ \d f m -> pmatch m $
+    \case
+      PDJust x -> f #$ pfield @"_0" # x
+      _ -> d
+
+{- | Construct a 'PDJust' value.
+
+@since 2.1.1
+-}
+pdjust ::
+  forall (a :: PType) (s :: S).
+  PIsData a =>
+  Term s (a :--> PMaybeData a)
+pdjust = phoistAcyclic $
+  plam $
+    \x -> pcon $ PDJust $ pdcons @"_0" # pdata x #$ pdnil
+
+{- | Construct a 'PDNothing' value.
+
+@since 2.1.1
+-}
+pdnothing ::
+  forall (a :: PType) (s :: S).
+  Term s (PMaybeData a)
+pdnothing = phoistAcyclic $ pcon $ PDNothing pdnil
+
+{- | Construct a 'PMaybeData' given a 'PMaybe'. Could be useful if you want to
+"lift" from 'PMaybe' to 'Maybe'.
+
+@since 2.1.1
+-}
+pmaybeToMaybeData ::
+  forall (a :: PType) (s :: S).
+  PIsData a =>
+  Term s (PMaybe a :--> PMaybeData a)
+pmaybeToMaybeData = phoistAcyclic $
+  plam $ \t -> pmatch t $ \case
+    PNothing -> pcon $ PDNothing pdnil
+    PJust x -> pcon $ PDJust $ pdcons @"_0" # pdata x # pdnil
+
+{- | Extract the value stored in a 'PMaybeData' container. If there's no value,
+throw an error with the given message.
+
+@since 2.1.1
+-}
+passertPDJust ::
+  forall (a :: PType) (s :: S).
+  PIsData a =>
+  Term s (PString :--> PMaybeData a :--> a)
+passertPDJust = phoistAcyclic $
+  plam $ \emsg mv' -> pmatch mv' $ \case
+    PDJust ((pfield @"_0" #) -> v) -> v
+    _ -> ptraceError emsg
 
 -- Helpers
